requirements.txt
pytorch >= 1.2
opencv >= 3
scikit-image
numpy
tqdm
tensorboardX
